Add program exists check in hook scripts

+++ sitegen/templates/cron/service/cron_remove.sh
@@ -1,7 +1,6 @@
-CRON_CRONFILE={{ cron_conf_file }}
-CRON_PROJECTFILE={{ cron_project_file }}
+CRON_CRONFILE={{ cron_conffile }}
 
-# Check cron
-type -P cron &>/dev/null || { echo "I require cron but it's not installed.  Aborting." >&2; exit 1; }
-
-sudo rm -rf $CRON_CRONFILE
+if [ -f $CRON_CRONFILE ]; then
+    echo '  * Remove cron file '$CRON_CRONFILE'.'
+    sudo rm -rf $CRON_CRONFILE
+fi
